feat

+++ application/MVVM/ViewModel/Pages/StaffViewModel.cs
@@ -2,6 +2,7 @@ using System.Collections.ObjectModel;
 
 using application.Abstraction.Interfaces;
 using application.MVVM.Model;
+using application.Utilities;
 
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
@@ -36,6 +37,12 @@ public partial class StaffViewModel : ObservableObject
 	[RelayCommand]
 	public async Task DeleteStaffMemberAsync(Guid staffId)
 	{
+		if (!TableHelper.ShowConfirmationMessage(
+			"Вы действительно хотите удалить выбранный элемент?",
+			"Подтверждение удаления"
+			))
+			return;
+
 		var isDeleted = await _staffRepository.DeleteStaffMemberAsync(staffId);
 		if (isDeleted)
 		{
